+++ build-chemotion/scripts/backupScript.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+echo Running the backup script