+++ build-chemotion/scripts/cmd.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+case "$1" in
+    bash)
+        exec /bin/bash
+        ;;
+    *)
+        python3 /etc/scripts/CLI.py $@
+        ;;
+esac