+++ build-chemotion/scripts/stopELNscript.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+echo Running the stop ELN script
+echo Nothing to do.
+
+# ps -ef | grep puma | awk '{print $2}' | xargs kill