+++ client-chemotion/setup.sh
@@ -1,0 +1,12 @@
+#!/bin/bash
+set -e
+mkdir -p shared/eln/log
+mkdir -p shared/eln/tmp
+mkdir -p shared/eln/uploads
+mkdir -p shared/eln/public/packs
+mkdir -p shared/eln/public/assets
+mkdir -p shared/eln/public/images/ketcherails
+mkdir -p shared/eln/public/images/molecules
+mkdir -p shared/eln/public/images/samples
+mkdir -p shared/eln/config
+mkdir -p db-data