+++ build-chemotion/CHANGELOG.md
@@ -1,3 +1,10 @@
 # Changelog
 
-21-05-28: Upped version to 21.05-2, based on v0.9.0 of ELN.
+## 21-19-11
+
+-   Updated Chemotion ELN version to 1.0.3.
+-   new Docker middleware implemented (D0.1)
+
+## 21-05-28
+
+-   Upped version to 21.05-2, based on v0.9.0 of ELN.